Update for Dropwizard 0.7-SNAPSHOT
